ricksharebackend/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-crontab==0.7.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
sqlparse==0.5.1
